refactor: separate single window.rs file into appropriate modules
